v1.3.9: Refactor Phase 6 - CompareFilters Decomposition
